refactor of the writers. log to stderr
