+++ Cargo.toml
@@ -106,9 +106,6 @@ serde = { version = "1.0", features = [ "derive" ] }
 # json serializer
 serde_json = "1.0"
 
-# simple log framework
-simplelog = { version = "0.12", features = ["paris"] }
-
 # sqlx (async db)
 sqlx = { version = "0.6", features = [
     "runtime-tokio-rustls",
