+++ icann-rdap-cli/src/main.rs
@@ -2,7 +2,6 @@ use bootstrap::BootstrapType;
 use icann_rdap_common::check::CheckClass;
 use icann_rdap_common::client::create_client;
 use icann_rdap_common::client::ClientConfig;
-use query::PageWriter;
 use query::ProcessingParams;
 use std::str::FromStr;
 use tracing::error;
@@ -10,13 +9,15 @@ use tracing::info;
 #[cfg(debug_assertions)]
 use tracing::warn;
 use tracing_subscriber::filter::LevelFilter;
+use write::FmtWrite;
+use write::PagerWrite;
 
 use clap::{ArgGroup, Parser, ValueEnum};
 use error::CliError;
 use icann_rdap_client::query::qtype::QueryType;
 use icann_rdap_common::VERSION;
 use is_terminal::IsTerminal;
-use query::{BridgeWriter, OutputType};
+use query::OutputType;
 use reqwest::Client;
 use tokio::{join, task::spawn_blocking};
 
@@ -27,6 +28,7 @@ pub mod dirs;
 pub mod error;
 pub mod query;
 pub mod request;
+pub mod write;
 
 const BEFORE_LONG_HELP: &str = include_str!("before_long_help.txt");
 const AFTER_LONG_HELP: &str = include_str!("after_long_help.txt");
@@ -418,7 +420,10 @@ pub async fn main() -> anyhow::Result<()> {
     let rdap_client = create_client(&client_config);
     if let Ok(client) = rdap_client {
         if !use_pager {
-            tracing_subscriber::fmt().with_max_level(level).init();
+            tracing_subscriber::fmt()
+                .with_max_level(level)
+                .with_writer(std::io::stderr)
+                .init();
             let output = &mut std::io::stdout();
             let res1 = join!(exec(
                 cli.query_value,
@@ -430,13 +435,13 @@ pub async fn main() -> anyhow::Result<()> {
             res1.0?;
         } else {
             let pager = minus::Pager::new();
-            let output = BridgeWriter(pager.clone());
+            let output = FmtWrite(pager.clone());
             let pager2 = pager.clone();
 
             tracing_subscriber::fmt()
                 .with_max_level(level)
                 .with_writer(move || -> Box<dyn std::io::Write> {
-                    Box::new(PageWriter(pager2.clone()))
+                    Box::new(PagerWrite(pager2.clone()))
                 })
                 .init();
             let pager = pager.clone();
